+++ test/k8s/client/runner/base_test.exs
@@ -6,7 +6,7 @@ defmodule K8s.Client.Runner.BaseTest do
   alias K8s.Client
   alias K8s.Client.Runner.Base
   alias K8s.Client.DynamicHTTPProvider
-  import K8s.Test.KubeHelper
+  import K8s.Test.IntegrationHelper
 
   defmodule HTTPMock do
     @base_url "https://localhost:6443"
@@ -15,7 +15,9 @@ defmodule K8s.Client.Runner.BaseTest do
 
     def request(:get, @namespaced_url, _, _, _), do: render(nil)
 
-    def request(:post, @namespaced_url, body, _, _), do: render(body)
+    def request(:post, @namespaced_url, body, _, _) do
+      render(body)
+    end
 
     def request(:get, @namespaced_url <> "/test", _body, _headers, _opts) do
       render(nil)
@@ -42,7 +44,7 @@ defmodule K8s.Client.Runner.BaseTest do
           _body,
           _headers,
           ssl: _ssl,
-          params: %{labelSelector: "app=nginx"}
+          params: [labelSelector: "app=nginx"]
         ) do
       render(nil)
     end
@@ -60,72 +62,60 @@ defmodule K8s.Client.Runner.BaseTest do
 
   setup do
     DynamicHTTPProvider.register(self(), __MODULE__.HTTPMock)
-    conn = K8s.Conn.from_file("test/support/kube-config.yaml")
+    {:ok, conn} = K8s.Conn.from_file("test/support/kube-config.yaml")
 
     {:ok, conn: conn}
   end
 
-  describe "run/3" do
+  describe "run/2" do
     test "running an operation with a K8s.Selector set", %{conn: conn} do
       operation =
         Client.list("v1", :pods)
         |> K8s.Selector.label({"app", "nginx"})
 
-      assert {:ok, _} = Base.run(operation, conn)
+      assert {:ok, _} = Base.run(conn, operation)
     end
 
     test "running an operation without an HTTP body", %{conn: conn} do
-      operation = Client.get(make_namespace("test"))
-      assert {:ok, _} = Base.run(operation, conn)
+      operation = Client.get(build_namespace("test"))
+      assert {:ok, _} = Base.run(conn, operation)
     end
 
     test "running an operation with an HTTP body", %{conn: conn} do
-      operation = Client.create(make_namespace("test"))
-      assert {:ok, _} = Base.run(operation, conn)
+      operation = Client.create(build_namespace("test"))
+      assert {:ok, _} = Base.run(conn, operation)
     end
 
     test "running an operation with options", %{conn: conn} do
-      operation = Client.get(make_namespace("test-query-params"))
-      params = %{"watch" => "true"}
-      operation_w_params = Map.put(operation, :query_params, params)
+      operation = Client.get(build_namespace("test-query-params"))
+      operation_w_params = K8s.Operation.put_query_param(operation, :watch, true)
 
-      assert {:ok, ^params} = Base.run(operation_w_params, conn)
+      # TODO: this should assert that the params were actually added
+      assert {:ok, _} = Base.run(conn, operation_w_params)
     end
 
     test "supports subresource operations", %{conn: conn} do
       operation = Client.get("apps/v1", "deployments/status", name: "nginx", namespace: "default")
-      assert {:ok, _} = Base.run(operation, conn)
-    end
-
-    test "running an operation with a custom HTTP body", %{conn: conn} do
-      operation = Client.create(make_namespace("test"))
-      labels = %{"env" => "test"}
-      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
-
-      assert {:ok, body} = Base.run(operation, conn, body)
-
-      assert body ==
-               ~s({"apiVersion":"v1","kind":"Namespace","metadata":{"labels":{"env":"test"},"name":"test"}})
+      assert {:ok, _} = Base.run(conn, operation)
     end
   end
 
-  describe "run/4" do
-    test "[DEPRECATED] running an operation with a custom HTTP body and options", %{
-      conn: conn
-    } do
-      operation = Client.create(make_namespace("test"))
-      labels = %{"env" => "test"}
-      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
+  describe "run/3" do
+    test "passes http_opts to the HTTP Provider", %{conn: conn} do
+      operation = Client.create(build_namespace("http-opts"))
 
-      opts = [params: %{"watch" => "true"}]
-      assert {:ok, _} = Base.run(operation, conn, body, opts)
+      assert {:ok, body} = Base.run(conn, operation, stream_to: self())
+
+      # TODO: this should assert that the opts were actually received
+      assert body ==
+               ~s({"apiVersion":"v1","kind":"Namespace","metadata":{"name":"http-opts"}})
     end
   end
 
   describe "run" do
     test "request with HTTP 2xx response", %{conn: conn} do
       operation = Client.list("v1", "Namespace", [])
-      assert {:ok, _} = Base.run(operation, conn)
+      assert {:ok, _} = Base.run(conn, operation)
     end
 
     test "supports subresource operations with alternate `kind` HTTP bodies", %{conn: conn} do
@@ -148,7 +138,7 @@ defmodule K8s.Client.Runner.BaseTest do
       }
 
       operation = K8s.Client.create(pod, eviction)
-      assert {:ok, _} = Base.run(operation, conn)
+      assert {:ok, _} = Base.run(conn, operation)
     end
   end
 end