+++ test/k8s/client/runner/stream_test.exs
@@ -1,5 +1,5 @@
+# credo:disable-for-this-file
 defmodule K8s.Client.Runner.StreamTest do
-  # credo:disable-for-this-file
   use ExUnit.Case, async: true
   doctest K8s.Client.Runner.Stream
   doctest K8s.Client.Runner.Stream.ListRequest
@@ -10,39 +10,41 @@ defmodule K8s.Client.Runner.StreamTest do
     @base_url "https://localhost:6443"
     @namespaced_url @base_url <> "/api/v1/namespaces"
     import K8s.Test.HTTPHelper
-    import K8s.Test.KubeHelper
+    import K8s.Test.IntegrationHelper
 
     def request(:get, @namespaced_url <> "/stream-empty-test/services", _body, _headers, _opts) do
-      data = make_list([])
+      data = build_list([])
       render(data, 200)
     end
 
     def request(:get, @namespaced_url <> "/stream-failure-test/services", _, _, opts) do
       params = opts[:params]
-      page1_items = [make_service("foo", "stream-failure-test")]
+      page1_items = [build_service("foo", "stream-failure-test")]
       continue_token = "stream-failure-test"
 
       case params do
-        %{continue: nil, limit: _limit} ->
-          data = make_list(page1_items, continue_token)
-          render(data, 200)
+        [limit: 10, continue: nil, labelSelector: ""] ->
+          data = build_list(page1_items, continue_token)
+          render(data, 200, [{"Content-Type", "application/json"}])
 
-        %{continue: ^continue_token} ->
-          render(%{}, 404)
+        [limit: 10, continue: "stream-failure-test", labelSelector: ""] ->
+          render(%{"message" => "next page not found"}, 404, [
+            {"Content-Type", "application/json"}
+          ])
       end
     end
 
     def request(:get, @namespaced_url <> "/stream-runner-test/services", _, _, opts) do
       params = opts[:params]
-      page1_items = [make_service("foo", "stream-runner-test")]
-      page2_items = [make_service("bar", "stream-runner-test")]
-      page3_items = [make_service("qux", "stream-runner-test")]
+      page1_items = [build_service("foo", "stream-runner-test")]
+      page2_items = [build_service("bar", "stream-runner-test")]
+      page3_items = [build_service("qux", "stream-runner-test")]
 
       body =
         case params do
-          %{continue: nil, limit: _limit} -> make_list(page1_items, "start")
-          %{continue: "start"} -> make_list(page2_items, "end")
-          %{continue: "end"} -> make_list(page3_items)
+          [limit: 10, continue: nil, labelSelector: ""] -> build_list(page1_items, "start")
+          [limit: 10, continue: "start", labelSelector: ""] -> build_list(page2_items, "end")
+          [limit: 10, continue: "end", labelSelector: ""] -> build_list(page3_items)
         end
 
       render(body, 200)
@@ -51,14 +53,14 @@ defmodule K8s.Client.Runner.StreamTest do
 
   setup do
     DynamicHTTPProvider.register(self(), __MODULE__.HTTPMock)
-    {:ok, conn} = K8s.Conn.lookup(:test)
+    {:ok, conn} = K8s.Conn.from_file("test/support/kube-config.yaml")
     {:ok, %{conn: conn}}
   end
 
   describe "run/3" do
     test "when the initial request has no results", %{conn: conn} do
       operation = K8s.Client.list("v1", "Service", namespace: "stream-empty-test")
-      assert {:ok, stream} = Stream.run(operation, conn)
+      assert {:ok, stream} = Stream.run(conn, operation)
 
       services = Enum.into(stream, [])
       assert services == []
@@ -66,7 +68,7 @@ defmodule K8s.Client.Runner.StreamTest do
 
     test "puts error tuples into the stream when HTTP errors are encountered", %{conn: conn} do
       operation = K8s.Client.list("v1", "Service", namespace: "stream-failure-test")
-      assert {:ok, stream} = Stream.run(operation, conn)
+      assert {:ok, stream} = Stream.run(conn, operation)
 
       services = Enum.into(stream, [])
 
@@ -82,7 +84,7 @@ defmodule K8s.Client.Runner.StreamTest do
 
     test "returns an enumerable stream of k8s resources", %{conn: conn} do
       operation = K8s.Client.list("v1", "Service", namespace: "stream-runner-test")
-      assert {:ok, stream} = Stream.run(operation, conn)
+      assert {:ok, stream} = Stream.run(conn, operation)
 
       services = Enum.into(stream, [])