+++ test/k8s/conn/auth/token_test.exs
@@ -8,17 +8,16 @@ defmodule K8s.Conn.Auth.TokenTest do
   describe "create/2" do
     test "creates a Token struct from token data" do
       auth = %{"token" => "wee"}
-      assert %Token{token: token} = Token.create(auth, nil)
-      assert token
+      assert {:ok, %Token{token: "wee"}} = Token.create(auth, nil)
     end
   end
 
   test "creates http request signing options" do
-    config = Conn.from_file("test/support/kube-config.yaml", user: "token-user")
-    assert %Token{token: token} = config.auth
+    {:ok, conn = %Conn{auth: %Token{token: token}}} =
+      Conn.from_file("test/support/kube-config.yaml", user: "token-user")
 
     {:ok, %Conn.RequestOptions{headers: headers, ssl_options: ssl_options}} =
-      Conn.RequestOptions.generate(config.auth)
+      Conn.RequestOptions.generate(conn.auth)
 
     assert headers == [{"Authorization", "Bearer #{token}"}]
     assert ssl_options == []