+++ test/k8s/middleware/request/initialize_test.exs
@@ -2,7 +2,7 @@ defmodule K8s.Middleware.Request.InitializeTest do
   use ExUnit.Case, async: true
 
   test "initializes a request headers from K8s.Conn.RequestOptions" do
-    conn = K8s.Conn.from_file("./test/support/kube-config.yaml")
+    {:ok, conn} = K8s.Conn.from_file("./test/support/kube-config.yaml")
     request = %K8s.Middleware.Request{conn: conn}
 
     {:ok, %{headers: headers}} = K8s.Middleware.Request.Initialize.call(request)
@@ -11,7 +11,7 @@ defmodule K8s.Middleware.Request.InitializeTest do
   end
 
   test "initializes a HTTPoison options from K8s.Conn.RequestOptions" do
-    conn = K8s.Conn.from_file("./test/support/kube-config.yaml")
+    {:ok, conn} = K8s.Conn.from_file("./test/support/kube-config.yaml")
     request = %K8s.Middleware.Request{conn: conn}
 
     {:ok, %{opts: opts}} = K8s.Middleware.Request.Initialize.call(request)